Split data_api tests into multiple files (one file per contract)
